New DB table 'text2tagset_defaults'.  Major refactoring of project access and file listing.
The new DB table 'text2tagset_defaults' defines project-specific defaults for the 'text2tagset' table, replacing the global variable in project_specific_hacks.php.

Access to projects and their associated data (including file listings) is now bundled in a new class ProjectAccessor.  This is still work in progress.

On the JS front-end, the AJAX query 'listFiles' was replaced with 'getProjectsAndFiles', which retrieves a list of projects, its associated files and default tagsets, and (if admin) project-specific settings and user links.  This function provides the data in a lot more structured and sensible manner than before, reducing the amount of JS code needed to handle it.  Finally, the resulting project list is stored in a new global variable cora.projects, which is a first step to get most CorA-specific globals into a single global 'cora' variable, acting as a kind of namespace.