Admins can now also change tagset associations for existing files.
In the file view, admins could always view tagset associations for each file via the "Tagsets..." link.  Now the dialog also allows changing these associations.

For security reasons, removing tagset associations is denied when there are any existing annotations for that tagset and the given file.  Adding new associations and removing those that don't have annotations is allowed.

On the negative side: The client-side code still uses the old tagset table instead of the new MultiSelect, and the server-side code for this functionality is a bit messy.