Primitive bash script for JS/CSS compression.
Added a script with calls to Closure compiler and YUICompressor for compression of JS/CSS files.  This is only a rough sketch of what commands to call on which files (and in what order!) to produce working minified JSS/CS files.  Should be made sane and integrated into a proper deployment script at some point.