UI changes to the text preview pane.
The horizontal text preview pane is now fixed to the bottom of the browser window.  A bottom margin is dynamically set for the containing div, so that all of its contents can still be scrolled into view.

Furthermore, whenever input elements in the editor gain focus, they now check if they're behind the header or the text preview, and if they are, the window is scrolled to compensate.