Line breaks in the transcription now are line breaks in the editor; removed project-specific stuff when editing tokens.
When adding or editing tokens, line breaks in the transcription are now actual line breaks in the dialog window.  This both makes clearer where line breaks occur and reduces project-specific features in the GUI.

Project-specific commentary for editing tokens, as well as the client-side check for correct line break symbols, was also removed.  This stuff can be done individually in the respective edit scripts for each project.