Main stylesheet is now an SCSS file.
For now, both the SCSS and the compiled CSS file are included in this repository.  This should be changed when a common deployment process has been established.

In the SCSS, styles have been nested whenever possible, and many (but not yet all) colors have been made into variables.  Further refinements should be made to make the SCSS more readable and maintainable.