Browser back/forward buttons can now be used to navigate between texts.
Also, files in CorA can be linked directly to other users by appending ?fid=<fileid> to the URL.  This URL is also shown in the browser bar when opening a new file.

Currently, using back/forward to navigate when a file is opened that has unsaved changes can cause troubles: If the user aborts the action, the browser history will point to the wrong place.  This problem will likely be eliminated with planned changes to the save functionality.