Requesting lines now uses CoraRequest; proper error handling.
Requesting lines should now behave in a sane way when requests fail, retrying them and alerting the user.  Returning to a page that is already loaded now works in this situation.  Changing pages more rapidly than the server can respond also behaves better.

For example, when a user loses the connection, the web interface still remains functional, pages already in memory can still be viewed, and CorA recovers properly when the connection is restored.  All of these have not been the case prior to this commit.