Changed training process for PHP Tagger objects.
Retrieving all tokens from all files in a project at once proved to be using huge amounts of memory.

As a partial solution, the training process was changed to be called incrementally, on one text at a time, so as to reduce the peak memory usage, will retaining the original functionality.