Generate CSS links with cache buster directly from CMake
Instead of PHP dynamically generating links to the CSS files with cache
buster additions, let CMake create them automatically during build
time.

MD5 hashes based on the files are used for the cache buster string.

Usage of add_sources(...) has been abandoned in favor of direct
configure_file(...) calls, because this ensures a re-configuration of
the CMakeLists.txt when one of the source files changes, necessary for
correctly updating the cache buster string.