Use LAB.js and move initialization to main.js (resolves #15 and #22)
LAB.js is now used as a script loader to fetch other files; this is done
in main.js (or main.login.js when on the login screen), and all
initialization code (ondomready events) has been moved here as well.

Currently, this breaks compatibility with debug mode, and also, the
main.js file(s) are not currently minified.