Make onLocaleChange callbacks editor-specific
Callback functions that belong to editor objects need to be unregistered
when the editor closes, otherwise they cause an exception. Since most
callback functions that trigger on locale change belong to the editor
anyway, we just redefine the function to be editor-specific for now, so
we can simply clear out all callbacks when the editor closes.